+++ app/src/main/java/com/rikvanvelzen/codingtest/data/models/dto/CurrencyRatesDTO.kt
@@ -4,7 +4,7 @@ data class CurrencyRatesDTO(
 
         val base: String? = null,
         val date: String? = null,
-        val rates: Map<String, Float>? = null
+        val rates: Map<String, Double>? = null
 )
