+++ app/src/main/java/com/rikvanvelzen/codingtest/data/repositories/CurrencyRepository.kt
@@ -68,7 +68,7 @@ class CurrencyRepository(private val currencyService: CurrencyService) {
                 baseCurrencyName,
                 countryDataProvider.getCountryCodeFromName(baseCurrencyName),
                 countryDataProvider.getCountryFlagUrl(baseCurrencyAbbreviation),
-                1F
+                1.toDouble()
         ))
 
         ratesResponse.rates?.forEach { (abbreviation, rate) ->
